refactor(ui): extract useSaveShortcut + useUnsavedWarning
Three editor surfaces (AgentEditor, SkillEditor, SettingsPage's
OrchestratorPromptSection) carried identical Cmd/Ctrl+S and
beforeunload boilerplate. Extract into src/lib/use-editor-shortcuts.ts.
