+++ src/lib/use-editor-shortcuts.ts
@@ -1,0 +1,26 @@
+import { useEffect } from 'react';
+
+/** Fires `save` when the user presses Cmd/Ctrl+S anywhere in the window. */
+export function useSaveShortcut(save: () => void) {
+  useEffect(() => {
+    const handler = (e: KeyboardEvent) => {
+      if ((e.metaKey || e.ctrlKey) && e.key === 's') {
+        e.preventDefault();
+        save();
+      }
+    };
+    window.addEventListener('keydown', handler);
+    return () => window.removeEventListener('keydown', handler);
+  }, [save]);
+}
+
+/** Warns before unload when `isDirty` is true. */
+export function useUnsavedWarning(isDirty: boolean) {
+  useEffect(() => {
+    const handler = (e: BeforeUnloadEvent) => {
+      if (isDirty) e.preventDefault();
+    };
+    window.addEventListener('beforeunload', handler);
+    return () => window.removeEventListener('beforeunload', handler);
+  }, [isDirty]);
+}
